+++ src/main/java/com/hyoguoo/paymentplatform/order/presentation/OrderViewController.java
@@ -1,5 +1,6 @@
-package study.paymentintegrationserver.controller;
+package com.hyoguoo.paymentplatform.order.presentation;
 
+import com.hyoguoo.paymentplatform.order.presentation.port.OrderService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -8,29 +9,24 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import study.paymentintegrationserver.dto.order.OrderCancelRequest;
 import study.paymentintegrationserver.dto.order.OrderCancelResponse;
-import study.paymentintegrationserver.service.OrderFacadeService;
-import study.paymentintegrationserver.service.OrderService;
 
 @Controller
-@RequestMapping("/order")
 @RequiredArgsConstructor
 public class OrderViewController {
 
     private final OrderService orderService;
-    private final OrderFacadeService orderFacadeService;
 
-    @GetMapping("/{id}")
+    @GetMapping("/order/{id}")
     public String findOrder(@PathVariable("id") Long id, Model model) {
-        model.addAttribute("order", orderFacadeService.getOrderDetailsByIdAndUpdatePaymentInfo(id));
+        model.addAttribute("order", orderService.getOrderDetailsByIdAndUpdatePaymentInfo(id));
 
         return "order/order-detail";
     }
 
-    @GetMapping()
+    @GetMapping("/order")
     public String findAllOrders(Model model,
             @RequestParam(value = "page", defaultValue = "0") int page,
             @RequestParam(value = "size", defaultValue = "20") int size
@@ -40,9 +36,9 @@ public class OrderViewController {
         return "order/order-list";
     }
 
-    @PostMapping("/cancel")
+    @PostMapping("/order/cancel")
     public String cancelOrder(@ModelAttribute @Valid OrderCancelRequest orderCancelRequest) {
-        OrderCancelResponse orderCancelResponse = orderFacadeService.cancelOrder(
+        OrderCancelResponse orderCancelResponse = orderService.cancelOrder(
                 orderCancelRequest
         );
